+++ data/entity/gen_entity.go
@@ -30,8 +30,8 @@ type TypeEntity struct {
 	InternalID  uint32
 	DisplayName string
 	Name        string
-	Width  		float64
-	Height 		float64
+	Width  		float32
+	Height 		float32
 	Type     	string
 }
